#255 [REFACTOR] AlarmManager 컴포넌트에서 eventBridge와 alarmServer 관리하도록 리팩토링
