--- operation-external/src/main/java/org/sopt/makers/operation/client/alarm/AlarmSender.java
@@ -1,7 +1,0 @@
-package org.sopt.makers.operation.client.alarm;
-
-import org.sopt.makers.operation.client.alarm.dto.AlarmSenderRequest;
-
-public interface AlarmSender {
-	void send(AlarmSenderRequest request);
-}
