+++ operation-external/src/main/java/org/sopt/makers/operation/client/alarm/alarmServer/AlarmSender.java
@@ -1,0 +1,7 @@
+package org.sopt.makers.operation.client.alarm.alarmServer;
+
+import org.sopt.makers.operation.client.alarm.alarmServer.dto.AlarmSenderRequest;
+
+public interface AlarmSender {
+    void send(AlarmSenderRequest request);
+}
