+++ operation-external/src/main/java/org/sopt/makers/operation/client/alarm/eventBridgeServer/EventBridgeSender.java
@@ -1,0 +1,9 @@
+package org.sopt.makers.operation.client.alarm.eventBridgeServer;
+
+
+import org.sopt.makers.operation.client.alarm.eventBridgeServer.dto.EventBridgeSenderRequest;
+
+public interface EventBridgeSender {
+    void scheduleAlarm(EventBridgeSenderRequest request, String postDate, String postTime, Long alarmId);
+
+}
