+++ operation-external/src/main/java/org/sopt/makers/operation/client/alarm/eventBridgeServer/dto/EventBridgeSenderRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.operation.client.eventbridge.dto;
+package org.sopt.makers.operation.client.alarm.eventBridgeServer.dto;
 
 import java.util.List;
 import org.sopt.makers.operation.alarm.domain.Category;
