--- operation-external/src/main/java/org/sopt/makers/operation/client/eventbridge/EventBridgeSender.java
@@ -1,9 +1,0 @@
-package org.sopt.makers.operation.client.eventbridge;
-
-
-import org.sopt.makers.operation.client.eventbridge.dto.EventBridgeSenderRequest;
-
-public interface EventBridgeSender {
-    void scheduleAlarm(EventBridgeSenderRequest request, String postDate, String postTime, Long alarmId);
-
-}
